Complete Phases 4, 5, and 6 of the refactor plan, enhancing service interface architecture, extension methods, and configuration management
This commit finalizes the implementation of Phases 4, 5, and 6 of the refactor plan, significantly improving the project's architecture. Key achievements include:

- **Phase 4: Service Interface Architecture**:
  - Confirmed that all major Infrastructure services have corresponding Application layer interfaces.
  - Created the `IPermissionHelper` interface for enhanced permission management.
  - Removed direct references to Infrastructure implementations from the UI layer, ensuring compliance with Clean Architecture principles.

- **Phase 5: Extension Methods Optimization**:
  - Evaluated and relocated `IdentityResultExtensions` to the Application layer, aligning with its semantic purpose.
  - Cleaned up redundant using directives, improving code quality and maintainability.

- **Phase 6: Configuration Management Optimization**:
  - Established a structured configuration management system using the `IAISettings` interface and the IOptions pattern.
  - Removed direct dependencies on `IConfiguration` from the UI layer, promoting better separation of concerns.
  - Ensured all configuration accesses are strongly typed, enhancing safety and reducing errors.

These changes align with ongoing efforts to enhance code organization and adherence to architectural guidelines, ensuring a more maintainable and scalable codebase.
