+++ src/Application/Common/Extensions/IdentityResultExtensions.cs
@@ -1,7 +1,9 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace CleanArchitecture.Blazor.Infrastructure.Extensions;
+using Microsoft.AspNetCore.Identity;
+
+namespace CleanArchitecture.Blazor.Application.Common.Extensions;
 
 public static class IdentityResultExtensions
 {
@@ -11,4 +13,4 @@ public static class IdentityResultExtensions
             ? Result.Success()
             : Result.Failure(result.Errors.Select(e => e.Description).ToArray());
     }
-}
+}
