+++ src/Server.UI/Pages/Identity/Roles/Roles.razor
@@ -189,7 +189,6 @@
     private bool _isLoading;
     private bool _isUploading;
     private bool _isExporting;
-    private PermissionHelper _permissionHelper = null!;
     private RolePermissionAssignmentService _permissionAssignmentService = null!;
     private RolesAccessRights _accessRights = new();
 
@@ -206,7 +205,6 @@
     private void InitializeServices()
     {
         _roleManager = ScopedServices.GetRequiredService<RoleManager<ApplicationRole>>();
-        _permissionHelper = ScopedServices.GetRequiredService<PermissionHelper>();
         _permissionAssignmentService = ScopedServices.GetRequiredService<RolePermissionAssignmentService>();
     }
