+++ src/Server.UI/Services/Identity/RolePermissionAssignmentService.cs
@@ -1,5 +1,6 @@
 ﻿using System.Security.Claims;
 using CleanArchitecture.Blazor.Application.Common.ExceptionHandlers;
+using CleanArchitecture.Blazor.Application.Common.Interfaces;
 using CleanArchitecture.Blazor.Application.Common.Interfaces.Identity;
 using CleanArchitecture.Blazor.Application.Common.Security;
 using CleanArchitecture.Blazor.Domain.Identity;
@@ -12,13 +13,13 @@ namespace CleanArchitecture.Blazor.Server.UI.Services.Identity;
 public class RolePermissionAssignmentService : IPermissionAssignmentService
 {
     private readonly RoleManager<ApplicationRole> _roleManager;
-    private readonly PermissionHelper _permissionHelper;
+    private readonly IPermissionHelper _permissionHelper;
     private readonly IFusionCache _cache;
     private readonly ILogger<RolePermissionAssignmentService> _logger;
     private const string CacheKeyPrefix = "get-claims-by-";
 
     public RolePermissionAssignmentService(RoleManager<ApplicationRole> roleManager,
-        PermissionHelper permissionHelper,
+        IPermissionHelper permissionHelper,
         IFusionCache cache,
         ILogger<RolePermissionAssignmentService> logger)
     {
