+++ src/Server.UI/_Imports.razor
@@ -29,7 +29,6 @@
 @using CleanArchitecture.Blazor.Application.Common.Models
 @using CleanArchitecture.Blazor.Application.Common.Security;
 @using CleanArchitecture.Blazor.Application.Common.Constants
-@using CleanArchitecture.Blazor.Application.Common.Extensions
 @using CleanArchitecture.Blazor.Server.UI.Models
 @using CleanArchitecture.Blazor.Server.UI
 @using CleanArchitecture.Blazor.Server.UI.Shared
@@ -52,7 +51,6 @@
 @inject IApplicationSettings ApplicationSettings
 @inject ISnackbar Snackbar
 @inject IDialogService DialogService
-@inject IConfiguration Config
 @inject IAuthorizationService AuthService 
 @inject IValidationService Validator
 @inject IJSRuntime JS
@@ -60,4 +58,5 @@
 @inject NavigationManager Navigation
 @inject DialogServiceHelper DialogServiceHelper
 @inject IPermissionService PermissionService
+@inject IPermissionHelper PermissionHelper
 @inject IMapper Mapper
